requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-multipart
pyyaml
requests
python-dotenv
